pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nl2protocol"
version = "0.3.0"
description = "Convert natural language to Opentrons robot protocols using Claude LLM"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["opentrons", "robotics", "llm", "protocol", "laboratory", "automation", "claude", "anthropic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "pydantic>=2.0.0",
    "anthropic>=0.40.0",
    "opentrons>=8.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "hypothesis>=6.0.0",
    "mutmut>=3.0.0",
]
pdf = [
    "pypdf>=4.0.0",
]

[project.scripts]
nl2protocol = "nl2protocol.cli:main"

[project.urls]
Homepage = "https://github.com/aryan01104/nl2protocol"
Repository = "https://github.com/aryan01104/nl2protocol"

[tool.hatch.build.targets.wheel]
packages = ["nl2protocol"]

[tool.hatch.build.targets.sdist]
include = [
    "nl2protocol/",
    "config_examples/",
    "robot_config.example.json",
    "README.md",
    "LICENSE",
]
